Build improvements * Cleaning up PanBase for easier setup and build purposes 	* Moving `PanBase` to its own file 	* Fix tests * Proper version (i.e. `pocs.__version` works)
