+++ pocs/version.py
@@ -2,4 +2,4 @@ major = 0
 minor = 5
 patch = 1
 
-version = '{}.{}.{}'.format(major, minor, patch)
+__version__ = '{}.{}.{}'.format(major, minor, patch)
